requirements.txt
numpy
pandas
scipy
matplotlib
pillow